Initial docker-compose rig, Express, and TypeORM.
Borrowed and tweaked and existing docker-compose rig to get up and
running. This is my first time using TypeORM, been wanting to check it
out for awhile. Very easy and fun to use!

The trickiest part was getting the process to wait for the database
connection, the MariaDB image seems to take longer than usual to be
ready.